modified to run parallelizable index procedures against index partitions in parallel.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@495 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/java/com/bigdata/rdf/inf/SPOAssertionBuffer.java
@@ -33,6 +33,7 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import com.bigdata.rdf.spo.ISPOAssertionBuffer;
 import com.bigdata.rdf.spo.ISPOFilter;
@@ -195,7 +196,7 @@ public class SPOAssertionBuffer extends AbstractSPOBuffer implements ISPOAsserti
                     stmts, numStmts), nwritten));
             
             // task will write justifications on the justifications index.
-            AtomicInteger nwrittenj = new AtomicInteger();
+            AtomicLong nwrittenj = new AtomicLong();
             tasks.add(new JustificationWriter(focusStore,
                     new JustificationArrayIterator(justifications,
                             numJustifications),nwrittenj));
